Moved prompt_response into its own module.
